+++ app/src/main/java/com/example/marvelheros/data/model/MarvelThumbnail.kt
@@ -1,0 +1,9 @@
+package com.example.marvelheros.data.model
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class MarvelThumbnail(
+    val path: String,
+    val extension: String
+)
