+++ app/src/main/java/com/example/marvelheros/domain/repository/HeroRepository.kt
@@ -1,0 +1,7 @@
+package com.example.marvelheros.domain.repository
+
+import com.example.marvelheros.domain.model.Hero
+
+interface HeroRepository {
+    suspend fun getHeroes(): List<Hero>
+}
