+++ app/src/main/java/com/example/marvelheros/domain/usecase/GetHeroesUseCase.kt
@@ -1,0 +1,12 @@
+package com.example.marvelheros.domain.usecase
+
+import com.example.marvelheros.data.repository.HeroRepository
+import com.example.marvelheros.domain.model.Hero
+import com.example.marvelheros.utils.MyResult
+import javax.inject.Inject
+
+class GetHeroesUseCase @Inject constructor(
+    private val repository: HeroRepository)
+{
+  suspend operator fun invoke(): MyResult<List<Hero>> = repository.getHeroes()
+}
