+++ app/src/main/java/com/example/marvelheros/ui/screen/MainScreen.kt
@@ -14,7 +14,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.example.marvelheros.data.model.Hero
+import com.example.marvelheros.domain.model.Hero
 import com.example.marvelheros.ui.components.FullScreenHeroDetails
 
 import com.example.marvelheros.ui.screen.HeroUiState
